183640408 tweak homepage layout (#614)
* Merge cluster performance stats and update styling

* Move related components to one place

* Use more appriopriate name

* Unify components names

* Reduce header size

* Add 'loading...' to cluster numbers component

* Reduce heading size

* Fix wrong data in vote distance methods
