+++ app/javascript/packs/clusters/cluster_numbers_component.js
