+++ app/javascript/packs/clusters/cluster_performance_stats_component.js
@@ -1,5 +1,5 @@
 import Vue from 'vue/dist/vue.esm';
-import DistanceStatsComponentTemplate from './distance_stats_component_template';
+import DistanceStatsComponentTemplate from './cluster_performance_stats_component_template';
 import TurbolinksAdapter from 'vue-turbolinks';
 import store from "../stores/main_store.js";
 import ActionCableVue from "actioncable-vue";
