+++ app/javascript/packs/epochs/epoch_stats_component.js
