+++ app/javascript/packs/epochs/epoch_stats_component_template.vue
