+++ app/javascript/packs/navigations/network_buttons_template.vue
