+++ app/javascript/packs/ping_things/bubble_chart.js
