+++ app/javascript/packs/ping_things/index.js
