+++ app/javascript/packs/ping_things/index_template.vue
