+++ app/javascript/packs/ping_things/ping_thing_table.vue
