+++ app/javascript/packs/ping_things/stats_bar.js
