+++ app/javascript/packs/ping_things/stats_chart.js
