+++ app/views/layouts/_header.html.erb
@@ -1,4 +1,4 @@
-<%= javascript_pack_tag 'navigation/network_buttons' %>
+<%= javascript_pack_tag 'navigations/network_buttons' %>
 
 <header>
   <nav class="navbar navbar-expand-lg">
