+++ app/views/ping_things/index.html.erb
@@ -1,4 +1,4 @@
-<%= javascript_pack_tag 'ping_thing/index' %>
+<%= javascript_pack_tag 'ping_things/index' %>
 
 <div
   id='ping-thing-index-vue'
