+++ app/views/validators/_ping_stats_banner.html.erb
@@ -1,4 +1,4 @@
-<%= javascript_pack_tag 'ping_thing/stats_bar.js' %>
+<%= javascript_pack_tag 'ping_things/stats_bar.js' %>
 
 <div id="ping-thing-stats">
   <stats-bar network=<%= params[:network] %> title-visible="true"/>
