+++ app/views/validators/home_new.html.erb
@@ -1,12 +1,11 @@
 <%= javascript_pack_tag "validators/validators_map.js" %>
 <%= javascript_pack_tag 'transactions/transactions_stats_component' %>
 <%= javascript_pack_tag 'sol_prices/sol_price_component' %>
-<%= javascript_pack_tag 'epoch/epoch_stats_component' %>
+<%= javascript_pack_tag 'epochs/epoch_stats_component' %>
 <%= javascript_pack_tag 'validators/circulating_supply.js' %>
 <%= javascript_pack_tag 'leaders/leaders_stats_component.js' %>
-<%= javascript_pack_tag 'cluster_numbers/cluster_numbers_component.js' %>
-<%= javascript_pack_tag 'distance_stats/distance_stats_component.js' %>
-<%= javascript_pack_tag 'skipped_stats/skipped_stats_component.js' %>
+<%= javascript_pack_tag 'clusters/cluster_numbers_component.js' %>
+<%= javascript_pack_tag 'clusters/cluster_performance_stats_component.js' %>
 
 <% if @network == "mainnet" %>
   <%= render "validators/ping_stats_banner" %>
@@ -36,12 +35,6 @@
   </div>
 </section>
 
-<section>
-  <div id="distance-stats-component">
-</section>
-
-<section>
-  <div id="skipped-stats-component">
-</section>
+<div id="distance-stats-component"></div>
 
 <%= render 'validators/stake_pools_banner' %>
